+++ server/src/main/java/com/moovda_project/moovda/global/auth/userdetails/LoginDto.java
@@ -1,4 +1,4 @@
-package com.moovda_project.moovda.global.oauth2_jwt.userdetails;
+package com.moovda_project.moovda.global.auth.userdetails;
 
 import lombok.Getter;
